feat: add initial database migration with user, account, session, project, activity log, and git commit tables

+++ app/api/graphql/route.ts
@@ -27,7 +27,7 @@ const typeDefs = `
     projectPath: String!
     filePath: String!
     language: String!
-    timestamp: Float!
+    timestamp: BigInt!
     duration: Int!
   }
 
@@ -35,7 +35,7 @@ const typeDefs = `
     projectPath: String!
     filePath: String!
     language: String!
-    timestamp: Float!
+    timestamp: BigInt!
     duration: Int!
   }
